requirements.txt
Flask
qrcode
Pillow
nltk
gunicorn
authlib
Flask-Mail
python-dotenv
requests
deep-translator
langdetect
google-generativeai
